Additions and Rewrites (#11)
* Rewrites and additions

I added the `has_pending?` and `on_ready` that were requested as part of the previous attempt at a unified specification (or that's how I remember it).

`on_shutdown` is important for politely unwinding a process within a cluster, such as shutting down a Dyno in Heroku. I'm not sure servers should be forced to implement the feature, but it's nice to make it available for those that do (let me know what you think).

I also changed the `on_message` signature. I didn't understand what purpose `subject` was serving, so I removed it.

I allowed the `rack.upgrade?` to be missing as well as `nil` when the request isn't "upgradable".

I also reworded some stuff for you to consider (just see if they feel clearer, English is not my native language so maybe my suggestions aren't all that great).

* Clarify `on_ready`

* change `has_pending?` to `pending?`